Adding an OptionValueSet representation of sets of option.
This is a refactoring for making the configuration represented
as a dict of type/value pairs rather than as a command-line format string.

Added an OptionValueSet that stores name/value pairs, not a string.
It retains the ability to remember parts of configuration that it has been
passed, but does not understand.

The resulting config is sorted and normalized to 1 space between options.

Moved GetValue/ChangeValue from Option class to OptionValueSet class.

Also added a DummyOption class, to represent program-changed parameters.