Generate result pages in website.
This CL creates and populates an index page in /results that will auto-update
with the available content of the site (stored in /results/generated).

It adds a new tool "generate_pages" that will regenerate all data served
by the web site, and moves parts of the compare_html function into functions
on visual_metrics.py, where they can be tested.