Refactoring: Separate "codec" from "evaluation context"
Moved the "cache" member from "codec" to a new class "context".
This means that the "codec" class is now pure read-only.
Moved the "score" function from "codec" to a new class "optimizer",
which is a subclass of "context". Gathered everything that has to do
with scoring and picking a "best" encoding into the "optimizer" class.

This change had knock-on changes in most other files.
Also fixed at least one bug found by testing SomeUntriedVariants.