Added a try-otherwise-useful encoder algorithm to optimizer.
This algorithm will look through the list of other encodings done with
the same encoder as the currently best one, and see if there are encoders
that give better encodings. - the function WorksBetterOnSomeOtherClip.

The CL also removes the "synthesize cliptime" hack; to remove all
results that don't contain cliptime, do this:

  grep -L 'cliptime' workdir/*/*/*/*.result | xargs rm

The result seems to be a large improvement in number of probes that result
in an improvement when running randomly_improve_encoding (current run: 70
improvements to 8 not improved, vs. approx equal numbers before).