Make RT mode encoding for VP8 be stable.
In mode --rt, and with --cpu-used set to >= 0, the vp8 encoder will
create different files depending on the machine load. Setting --cpu-used
to a negative number alleviates this.

This CL also contains a tool for verifying that a configuration is
stable, as well as a fix to the hashname verification that checks that
the configuration is unchanged by the ConfigurationFixups call.