Conversion of results display to a different method.
This change passes the encoding results from the generation side to the
display side as JSON files. All usage of gviz code on the generation side
is no longer needed.

Note: This is a transition CL - the generation side will be able to generate
both the old way and the new way until the new way works better than the old
way.