A cross performance table that uses single encoders.
Removed the "old/new format" parameter from a few functions in
visual_metrics since all the uses for that had disappeared.
Removed the "compare_html" tool.

Added displaying the new graphs, with various knock-on changes.
Side effort: added an update to the "fetch-files" script.

Also flipped the direction of the summary tables, so that baseline is
on the left.