Adding two HEVC codecs, reporting on one of them.
This PR adds the X265 and HEVC implementations of the HEVC (H.265) codec.

Neither codec has parameters that can vary yet.
Reporting exposes only the X265 codec so far.

Note - there is a bug in ffmpeg's decoding of HEVC files that sometimes
produces the wrong number of frames - as a stopgap measure, the decoder
from the HEVC Reference Model is used.