Adds running tools without all YUV files in a fileset.
This extends FilesAndRates to optionally check for whether files exist,
and omit files from the fileset that do not exist on the disk. This allows
testing the code without having a copy of all the files.

Also starts updating the pylint checker to a new version of pylint (1.1),
which does a lot more checking. Not all errors suppressed yet.