Fixes qmin / qmax relationship on the mjpeg codec.
This change ensures that qmax is always >= qmin when both are
specified for the mjpeg codec. It also ensures that qmax is given
if qmin > 31 (seems to be the hidden default for qmax).

It also adds a --dry-run flag to the randomly_improve_encoding script,
and the ability to randomly improve only for a single codec.