Add "--threads" parameter to x264 codec.
This adds a parameter for controlling the number of threads to
the x264 codec, removing the previously hardcoded default.
NOTE: WILL INVALIDATE STORED ENCODES.

It also adds a new tool "vary_parameter" that lets one run encodes
for all possible values of a parameter, so that response curves
can be seen.

It also fixes a minor Lint error in openh264.py.