Refactoring the disk cache to have directory as an init parameter
Running calculations of scores & best scores over several directories
turned out to be complex if directory was passed as a call parameter.

This refactoring makes it part of the init parameters of the optimizer,
the context and the cache.

This means that one has to have functions for "adopting" an encoding
into the other context.

Fixes #77